+++ metrox-android/src/main/kotlin/dev/zacsweers/metrox/android/BroadcastReceiverKey.kt
@@ -7,6 +7,8 @@ import dev.zacsweers.metro.MapKey
 import kotlin.reflect.KClass
 
 /** A [MapKey] annotation for binding a BroadcastReceiver in a multibinding map. */
-@MapKey
+@MapKey(implicitClassKey = true)
 @Target(AnnotationTarget.CLASS)
-public annotation class BroadcastReceiverKey(val value: KClass<out BroadcastReceiver>)
+public annotation class BroadcastReceiverKey(
+  val value: KClass<out BroadcastReceiver> = Nothing::class
+)