+++ metrox-viewmodel/src/commonMain/kotlin/dev/zacsweers/metrox/viewmodel/ViewModelKeys.kt
@@ -8,7 +8,7 @@ import dev.zacsweers.metro.MapKey
 import kotlin.reflect.KClass
 
 /** A [MapKey] annotation for binding ViewModels in a multibinding map. */
-@MapKey
+@MapKey(implicitClassKey = true)
 @Target(
   AnnotationTarget.FUNCTION,
   AnnotationTarget.FIELD,
@@ -18,13 +18,13 @@ import kotlin.reflect.KClass
   AnnotationTarget.TYPE,
 )
 @Retention(AnnotationRetention.RUNTIME)
-public annotation class ViewModelKey(val value: KClass<out ViewModel>)
+public annotation class ViewModelKey(val value: KClass<out ViewModel> = Nothing::class)
 
 /**
  * A [MapKey] annotation for binding [assisted ViewModel factories][ViewModelAssistedFactory] in a
  * multibinding map.
  */
-@MapKey
+@MapKey(implicitClassKey = true)
 @Target(
   AnnotationTarget.FUNCTION,
   AnnotationTarget.FIELD,
@@ -34,13 +34,15 @@ public annotation class ViewModelKey(val value: KClass<out ViewModel>)
   AnnotationTarget.TYPE,
 )
 @Retention(AnnotationRetention.RUNTIME)
-public annotation class ViewModelAssistedFactoryKey(val value: KClass<out ViewModel>)
+public annotation class ViewModelAssistedFactoryKey(
+  val value: KClass<out ViewModel> = Nothing::class
+)
 
 /**
  * A [MapKey] annotation for binding
  * [manually assisted ViewModel factories][ManualViewModelAssistedFactory] in a multibinding map.
  */
-@MapKey
+@MapKey(implicitClassKey = true)
 @Target(
   AnnotationTarget.FUNCTION,
   AnnotationTarget.FIELD,
@@ -51,7 +53,7 @@ public annotation class ViewModelAssistedFactoryKey(val value: KClass<out ViewMo
 )
 @Retention(AnnotationRetention.RUNTIME)
 public annotation class ManualViewModelAssistedFactoryKey(
-  val value: KClass<out ManualViewModelAssistedFactory>
+  val value: KClass<out ManualViewModelAssistedFactory> = Nothing::class
 )
 
 /**